Weapon/Flavor UI
Homing bullets now use "GameObject.FindObjectsWithTag" instead of looping through every object in the scene.
Fixed the constant warning about not implementing the "Awake()" function in the Strawberry.cs script.
Fixed spelling of Strawberry and Pomegranate.

You now also have a visual indicator of which weapon you have equipped in the bottom right corner.
This also means that every weapon has an icon, a description and some stats to go along with it.
There is now also a popup for when you obtain a weapon (TIP: Press U to obtain a new weapon!)